Move `Callbacks` to `connlib-client-shared`
