+++ rust/connlib/clients/apple/src/lib.rs
@@ -7,10 +7,9 @@ mod tun;
 use anyhow::Result;
 use backoff::ExponentialBackoffBuilder;
 use connlib_client_shared::{
-    callbacks::ResourceDescription, keypair, Callbacks, ConnectArgs, Error, LoginUrl, Session,
-    V4RouteList, V6RouteList,
+    keypair, Callbacks, ConnectArgs, Error, LoginUrl, Session, V4RouteList, V6RouteList,
 };
-use connlib_shared::get_user_agent;
+use connlib_shared::{callbacks::ResourceDescription, get_user_agent};
 use ip_network::{Ipv4Network, Ipv6Network};
 use phoenix_channel::PhoenixChannel;
 use secrecy::{Secret, SecretString};
