+++ rust/connlib/shared/src/lib.rs
@@ -12,7 +12,6 @@ pub mod proptest;
 
 pub use boringtun::x25519::PublicKey;
 pub use boringtun::x25519::StaticSecret;
-pub use callbacks::Callbacks;
 pub use error::ConnlibError as Error;
 pub use error::Result;
 pub use phoenix_channel::{LoginUrl, LoginUrlError};
